+++ packages/xchain-bitcoin/src/const.ts
@@ -58,7 +58,7 @@ export const blockstreamExplorerProviders = {
 
 const testnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  'PLACEHOLDER_APIKEY',
+  process.env['SOCHAIN_API_KEY'] || '',
   BTCChain,
   AssetBTC,
   8,
@@ -66,7 +66,7 @@ const testnetSochainProvider = new SochainProvider(
 )
 const mainnetSochainProvider = new SochainProvider(
   'https://sochain.com/api/v3',
-  'PLACEHOLDER_APIKEY',
+  process.env['SOCHAIN_API_KEY'] || '',
   BTCChain,
   AssetBTC,
   8,
@@ -103,6 +103,7 @@ const testnetBlockcypherProvider = new BlockcypherProvider(
   AssetBTC,
   8,
   BlockcypherNetwork.BTCTEST,
+  process.env['BlOCKCYPHER_API_KEY'] || '',
 )
 
 const mainnetBlockcypherProvider = new BlockcypherProvider(
@@ -111,6 +112,7 @@ const mainnetBlockcypherProvider = new BlockcypherProvider(
   AssetBTC,
   8,
   BlockcypherNetwork.BTC,
+  process.env['BlOCKCYPHER_API_KEY'] || '',
 )
 export const BlockcypherDataProviders: UtxoOnlineDataProviders = {
   [Network.Testnet]: testnetBlockcypherProvider,