+++ packages/xchain-cosmos/__tests__/sdk-client.test.ts
@@ -247,6 +247,7 @@ describe('SDK Client Test', () => {
           tx: {
             body: {
               messages: [msgSend],
+              memo: '',
             },
           },
           timestamp: '2020-09-25T06:09:15Z',
@@ -297,6 +298,7 @@ describe('SDK Client Test', () => {
           tx: {
             body: {
               messages: [msgSend2],
+              memo: '',
             },
           },
           timestamp: '2020-09-25T06:09:15Z',
@@ -383,6 +385,7 @@ describe('SDK Client Test', () => {
         tx: {
           body: {
             messages: [msgSend],
+            memo: '',
           },
         },
         timestamp: '2020-09-25T06:09:15Z',
@@ -413,6 +416,7 @@ describe('SDK Client Test', () => {
       tx: {
         body: {
           messages: [msgSend2],
+          memo: '',
         },
       },
       timestamp: '2020-09-25T06:09:15Z',