+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.22.2 (2023-07-10)
+
+## Update
+
+- added process.env[apikey] config as default option to provider creation
+
 # v0.22.1 (2023-05-18)
 
 ## Add