+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,16 @@
+# v0.28.5 (2023-07-18)
+
+## Update
+
+- fixed amount returned from getTransactionData()
+- Bump comsos dep
+
+# v0.28.4 (2023-07-16)
+
+## Update
+
+- Refactored getTransactionData to use new cosmos memo type and bumped cosmos dep
+
 # v0.28.3 (2023-06-30)
 
 ## Update